refactor: move markdown to application
created a core folder
