+++ frontend/app_flowy/lib/startup/startup.dart
@@ -4,7 +4,7 @@ import 'package:app_flowy/startup/tasks/prelude.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
-import 'package:app_flowy/workspace/infrastructure/deps_resolver.dart';
+import 'package:app_flowy/startup/deps_resolver.dart';
 import 'package:app_flowy/user/infrastructure/deps_resolver.dart';
 import 'package:flowy_sdk/flowy_sdk.dart';
