+++ frontend/app_flowy/lib/workspace/application/doc/share_bloc.dart
@@ -1,5 +1,5 @@
 import 'package:app_flowy/workspace/application/doc/share_service.dart';
-import 'package:app_flowy/workspace/infrastructure/markdown/delta_markdown.dart';
+import 'package:app_flowy/workspace/application/markdown/delta_markdown.dart';
 import 'package:flowy_sdk/protobuf/flowy-folder-data-model/share.pb.dart';
 import 'package:flowy_sdk/protobuf/flowy-folder-data-model/view.pb.dart';
 import 'package:flowy_sdk/protobuf/flowy-error/errors.pb.dart';
