+++ frontend/app_flowy/lib/workspace/application/view/view_listener.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 import 'dart:typed_data';
-import 'package:app_flowy/workspace/infrastructure/repos/helper.dart';
+import 'package:app_flowy/core/helper.dart';
 import 'package:dartz/dartz.dart';
 import 'package:flowy_sdk/protobuf/dart-notify/subject.pb.dart';
 import 'package:flowy_sdk/protobuf/flowy-folder-data-model/view.pb.dart';
